pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.2.1",
    "fastapi>=0.109.0",
    "openai>=1.35.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "supabase>=2.3.4",
    "tiktoken>=0.5.0",
    "uvicorn[standard]>=0.27.0",
]
